requirements.txt
# RAG-Hybrid System Dependencies
# Core API
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Authentication
bcrypt>=4.0.0
PyJWT>=2.8.0

# RAG & LLM
langchain>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0

# AI Providers
anthropic>=0.50.0
openai>=1.0.0

# Frontend
streamlit>=1.35.0

# Utilities
tiktoken>=0.7.0
numpy>=2.0.0
pandas>=2.0.0

# PDF Support
pypdf>=4.0.0
